# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EscapeArrow API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/start": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Start",
        "operationId": "start_api_auth_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routes__auth__Start"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Verify",
        "operationId": "verify_api_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/google/callback": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Google Callback",
        "operationId": "google_callback_api_auth_google_callback_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/poll": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Google Poll",
        "operationId": "google_poll_api_auth_google_poll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GooglePoll"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/contact/start": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Contact Start",
        "operationId": "contact_start_api_auth_contact_start_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkContact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/contact/verify": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Contact Verify",
        "operationId": "contact_verify_api_auth_contact_verify_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GooglePoll": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "state",
          "key"
        ],
        "title": "GooglePoll"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkContact": {
        "properties": {
          "contact": {
            "type": "string",
            "title": "Contact"
          }
        },
        "type": "object",
        "required": [
          "contact"
        ],
        "title": "LinkContact"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verify": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "title": "Challenge Id"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{4}$",
            "title": "Code"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "challenge_id",
          "code"
        ],
        "title": "Verify"
      },
      "app__routes__auth__Start": {
        "properties": {
          "purpose": {
            "type": "string",
            "title": "Purpose",
            "default": "login"
          },
          "contact": {
            "type": "string",
            "title": "Contact"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "title": "Name",
            "default": ""
          },
          "username": {
            "type": "string",
            "maxLength": 24,
            "title": "Username",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "type": "object",
        "required": [
          "contact"
        ],
        "title": "Start"
      },
...
    }
  }
}